pyproject.toml
[tool.poetry]
name = "speech-to-text"
version = "0.1.0"
description = ""
authors = ["Vitor Wdson <vitor.wdson2@gmail.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
openai = "^0.27.2"
pyaudio = "^0.2.13"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
